script/tmkms-sgx/pyproject.toml
[tool.poetry]
name = "sgx"
version = "0.1.0"
description = ""
authors = ["allthatjazzleo <pangleo1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
azure-identity = "^1.5.0"
azure-keyvault-secrets = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
